requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-ollama
langchain-openai
pydantic
pandas
openpyxl
sqlalchemy
python-dotenv
mysql-connector-python
requests
json-repair
fpdf2
